CHG: ID of an agent moved into ShadowAgent (from AbstractAgent)
Previously and since I was single machine, I saw the entire world
with all agents (and thus ShadowAgents), and could identify them
easily with just pointers on them... but when their copies move
around the MPI-world and ID is the only reliable way to identify
them, I had to move the ID to (already) part of the ShadowAgent.
It actually makes sense: ID+agentType+Geometry(incl. AABB) is
now the minimal representation of an agent.